send a letter to email when creating an order and redirect to another page (to the payment page in the future)

+++ next-pizza-api/src/main/java/ru/alex/nextpizzaapi/database/entity/Order.java
@@ -28,7 +28,7 @@ public class Order extends AuditableEntity {
     private String token;
 
     @Column(name = "total_amount")
-    private Float totalAmount;
+    private Integer totalAmount;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status")
